+++ packages/tennis-tournaments/.gitignore
@@ -3,3 +3,4 @@ dist
 .DS_Store
 
 .vercel
+.solid
