+++ packages/tennis-tournaments/package.json
@@ -1,7 +1,7 @@
 {
   "name": "tennis-tournaments",
   "version": "0.0.1",
-  "description": "",
+  "type": "module",
   "scripts": {
     "dev": "solid-start dev",
     "build": "solid-start build",
@@ -9,16 +9,22 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "solid-devtools": "^0.27.3",
+    "@types/node": "^18.11.18",
+    "esbuild": "^0.14.54",
+    "postcss": "^8.4.21",
+    "solid-start-node": "^0.2.19",
     "typescript": "^5.1.3",
     "vite": "^4.3.9",
-    "vite-plugin-solid": "^2.7.0",
     "vite-plugin-solid-svg": "0.6.1"
   },
   "dependencies": {
-    "@solidjs/meta": "0.28.5",
-    "@solidjs/router": "0.8.2",
+    "@solidjs/meta": "^0.28.2",
+    "@solidjs/router": "^0.8.2",
     "solid-js": "^1.7.6",
-    "solid-start": "0.2.26"
+    "solid-start": "^0.2.26",
+    "undici": "^5.15.1"
+  },
+  "engines": {
+    "node": ">=16.8"
   }
 }
