--- packages/tennis-tournaments/src/App.module.css
@@ -1,4 +1,0 @@
-.App {
-  display: flex;
-  flex-direction: column;
-}
