+++ packages/tennis-tournaments/src/components/Loading.tsx
@@ -1,6 +1,6 @@
 import { Component } from "solid-js";
 import styles from "./Loading.module.css";
-import LoadingIcon from "./Icons/Loading.svg";
+import LoadingIcon from "../Icons/Loading.svg";
 
 const Loading: Component = () => (
   <div class={styles.Loading}>
