+++ packages/tennis-tournaments/src/entry-client.tsx
@@ -1,0 +1,3 @@
+import { mount, StartClient } from "solid-start/entry-client";
+
+mount(() => <StartClient />, document);
