+++ packages/tennis-tournaments/src/entry-server.tsx
@@ -1,0 +1,9 @@
+import {
+  createHandler,
+  renderAsync,
+  StartServer,
+} from "solid-start/entry-server";
+
+export default createHandler(
+  renderAsync((event) => <StartServer event={event} />)
+);
