+++ packages/tennis-tournaments/src/root.tsx
@@ -1,0 +1,35 @@
+// @refresh reload
+import { Routes } from "@solidjs/router";
+import { Suspense } from "solid-js";
+import {
+  Body,
+  FileRoutes,
+  Head,
+  Html,
+  Meta,
+  Scripts,
+  Title,
+} from "solid-start";
+import { ErrorBoundary } from "solid-start/error-boundary";
+
+export default function Root() {
+  return (
+    <Html lang="en">
+      <Head>
+        <Title>Tennis Tournaments</Title>
+        <Meta charset="utf-8" />
+        <Meta name="viewport" content="width=device-width, initial-scale=1" />
+      </Head>
+      <Body>
+        <Suspense>
+          <ErrorBoundary>
+            <Routes>
+              <FileRoutes />
+            </Routes>
+          </ErrorBoundary>
+        </Suspense>
+        <Scripts />
+      </Body>
+    </Html>
+  );
+}
