+++ packages/tennis-tournaments/src/routes/index.tsx
@@ -1,0 +1,3 @@
+import { TournamentsPage } from "../components/TournamentsPage";
+
+export default TournamentsPage;
