+++ packages/tennis-tournaments/src/routes/tournament/[tournamentId].tsx
@@ -1,0 +1,3 @@
+import { TournamentPage } from "../../components/TournamentPage";
+
+export default TournamentPage;
