+++ packages/tennis-tournaments/vite.config.ts
@@ -1,8 +1,6 @@
-/// <reference types="./vite-env-override.d.ts" />
 import { defineConfig, loadEnv } from "vite";
 import solidPlugin from "solid-start/vite";
 import solidSvg from "vite-plugin-solid-svg";
-// import devtools from 'solid-devtools/vite';
 
 export default defineConfig(({ mode }) => {
   const env = loadEnv(mode, process.cwd(), "");
@@ -14,15 +12,7 @@ export default defineConfig(({ mode }) => {
         RAPID_API_KEY: env.RAPID_API_KEY,
       },
     },
-    plugins: [
-      /* 
-        Uncomment the following line to enable solid-devtools.
-        For more info see https://github.com/thetarnav/solid-devtools/tree/main/packages/extension#readme
-      */
-      // devtools(),
-      solidPlugin(),
-      solidSvg(),
-    ],
+    plugins: [solidPlugin(), solidSvg()],
     server: {
       port: 3000,
     },
